Pagination convered to typescript
