--- src/Components/Pagination/constants.ts
@@ -1,12 +1,0 @@
-export interface PerPageOption {
-  title: string,
-  value: number
-}
-
-export const perPageOptions: PerPageOption[] = [
-  { title: '5', value: 5 },
-  { title: '10', value: 10 },
-  { title: '15', value: 15 },
-  { title: '20', value: 20 },
-  { title: '25', value: 25 },
-];
