+++ tsconfig.json
@@ -5,6 +5,8 @@
         "module": "es6",
         "target": "es5",
         "jsx": "react",
-        "allowJs": true
+        "allowJs": true,
+        "moduleResolution": "node",
+        "allowSyntheticDefaultImports": true
     }
 }
